+++ sam/models/__init__.py
@@ -10,8 +10,10 @@ from .keras_templates import (  # noqa: F401
     create_keras_quantile_mlp,
     create_keras_quantile_rnn,
 )
-from .linear_quantile_regression import LinearQuantileRegression  # noqa: F401
+
+from .sam_shap_explainer import SamShapExplainer  # noqa: F401
+from .linear_model import LinearQuantileRegression  # noqa: F401
 from .base_model import BaseTimeseriesRegressor  # noqa: F401
 from .spc_model import ConstantTimeseriesRegressor  # noqa: F401
 from .mlp_model import MLPTimeseriesRegressor  # noqa: F401
-from .sam_shap_explainer import SamShapExplainer  # noqa: F401
+
