+++ sam/models/linear_model.py
